return_radar/accounts/views.py

from .serializers import *
from rest_framework_simplejwt.views import TokenObtainPairView
from rest_framework import generics, status
from rest_framework.response import Response
from rest_framework.permissions import AllowAny
from django.contrib.auth import get_user_model
from .tasks import send_otp
from rest_framework_simplejwt.tokens import RefreshToken
from rest_framework.views import APIView
from django_otp.plugins.otp_email.models import EmailDevice
from django_otp.oath import TOTP
User = get_user_model()

class RegisterView(generics.CreateAPIView):
    permission_classes = (AllowAny,)
    serializer_class = UserSerializer
    queryset = get_user_model().objects.all()

class UserList(generics.ListAPIView):
    serializer_class = UserSerializer
    queryset = User.objects.all()
class VerifyOTPView(generics.CreateAPIView):
    permission_classes = (AllowAny,)
    serializer_class = UserSerializer

    def post(self, request, *args, **kwargs):
        email = request.data.get('email')
        otp = request.data.get('otp')
        device = EmailDevice.objects.filter(email=email).first()
        if not device or not device.verify_token(otp):
            return Response({'detail': 'Invalid OTP.'}, status=status.HTTP_400_BAD_REQUEST)
        serializer = self.get_serializer(data=request.data)
        serializer.is_valid(raise_exception=True)
        self.perform_create(serializer)
        return Response(serializer.data, status=status.HTTP_201_CREATED)
class CustomTokenObtainPairView(TokenObtainPairView):
    serializer_class = CustomTokenObtainPairSerializer